AD.3 retire claude backend compatibility line
