+++ boundaries/atm-daemon/daemon-inbox-export.toml
@@ -21,7 +21,7 @@ io_forbidden = ["sqlite", "socket_io", "process_spawn"]
 
 [dependencies]
 allowed_dependents = []
-allowed_dependencies = ["atm-core", "atm-storage-claude"]
+allowed_dependencies = ["atm-core"]
 forbidden_edges = ["atm -> atm-daemon", "atm-graft -> atm-daemon"]
 
 [references]
@@ -45,4 +45,4 @@ review_gates = ["no_public_impl", "no_public_constructor", "no_direct_mailbox_he
 
 [status]
 state = "concrete_landed"
-notes = ["atm-storage-claude owns the Claude export contract; daemon runtime now supplies the crate-root adapter implementation", "send and receive compatibility writes must stay behind this adapter rather than reaching mailbox helpers directly", "watcher/reconcile handling for ATM-authored compatibility projection updates must remain idempotent under this adapter boundary", "Yb Y.10 leaves this adapter available only for explicit repair/rebuild export and removes normal-runtime append fallback into this seam"]
+notes = ["daemon-local historical Claude compatibility helpers own the retired inbox export contract until reconcile/watch deletion lands", "send and receive compatibility writes must stay behind this adapter rather than reaching mailbox helpers directly", "watcher/reconcile handling for ATM-authored compatibility projection updates must remain idempotent under this adapter boundary", "Yb Y.10 leaves this adapter available only for explicit repair/rebuild export and removes normal-runtime append fallback into this seam"]
