+++ boundaries/atm-daemon/daemon-inbox-ingress.toml
@@ -21,7 +21,7 @@ io_forbidden = ["sqlite", "socket_io", "process_spawn"]
 
 [dependencies]
 allowed_dependents = []
-allowed_dependencies = ["atm-core", "atm-storage-claude"]
+allowed_dependencies = ["atm-core"]
 forbidden_edges = ["atm -> atm-daemon", "atm-graft -> atm-daemon"]
 
 [references]
@@ -43,4 +43,4 @@ review_gates = ["no_public_impl", "no_public_constructor", "no_direct_mailbox_he
 
 [status]
 state = "active"
-notes = ["atm-storage-claude owns the Claude ingress contract; daemon runtime now supplies the crate-root adapter implementation", "watcher-driven reconcile now routes through this boundary rather than directly to mailbox helpers"]
+notes = ["daemon-local historical Claude compatibility helpers own the retired inbox ingress contract until reconcile/watch deletion lands", "watcher-driven reconcile routes through this boundary rather than directly to mailbox helpers"]
