--- boundaries/atm-storage-claude/message-store.toml
@@ -1,47 +1,0 @@
-boundary_id = "BOUNDARY-ClaudeMessageStore"
-owner_package = "atm-storage-claude"
-owner_crate_path = "atm_storage_claude"
-name = "MessageStoreImplementation"
-
-[public]
-trait = "MessageStore"
-notes = "Claude inbox backend implementation of the shared message-store contract"
-
-[implementation]
-type = "ClaudeStorageBackend"
-module = "atm_storage_claude"
-visibility = "private"
-constructor = "private"
-
-[composition]
-roots = ["atm_storage_claude::ClaudeStorageBackend::new"]
-
-[ownership]
-io_owns = ["claude_inbox_json_storage", "compat_source_projection_loading"]
-io_forbidden = ["socket_io", "process_spawn"]
-
-[dependencies]
-allowed_dependents = ["atm-runtime", "atm-daemon"]
-allowed_dependencies = ["atm-storage"]
-forbidden_edges = ["atm-storage-claude -> atm-core"]
-
-[references]
-scope = "outside_owner_crate"
-forbidden = []
-
-[contracts]
-request_types = ["Message", "MessageKey", "MessageQuery"]
-response_types = ["Message", "Vec<Message>"]
-error_types = ["AtmError"]
-
-[testing]
-allowed_test_double_paths = []
-forbidden_test_bypasses = []
-
-[enforcement]
-lint_rules = ["LINT-BOUNDARY-CLAUDE-MESSAGESTORE-DEPENDENCY"]
-review_gates = ["no_atm_core_dependency"]
-
-[status]
-state = "concrete_landed"
-notes = ["AC.2 created the Claude inbox backend implementation behind atm-storage::MessageStore"]
