+++ crates/atm-daemon/Cargo.toml
@@ -13,7 +13,6 @@ homepage.workspace = true
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.3" }
 atm-runtime = { path = "../atm-runtime", version = "1.2.3" }
 atm-storage = { path = "../atm-storage", version = "1.2.3" }
-atm-storage-claude = { path = "../atm-storage-claude", version = "1.2.3" }
 arc-swap.workspace = true
 chrono = { version = "0.4", features = ["serde"] }
 fs2 = "0.4"
