+++ crates/atm-daemon/src/lib.rs
@@ -8,6 +8,7 @@
 mod active_connection_registry;
 mod advisory_runtime;
 mod boundary_adapters;
+mod claude_compat;
 pub(crate) mod composition;
 mod daemon_runtime_observability;
 mod direct_boundaries;
