--- crates/atm-storage-claude/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "atm-storage-claude"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-description = "Claude-compatible mailbox and roster storage backend for atm-storage."
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-
-[dependencies]
-atm-storage = { path = "../atm-storage", version = "1.2.3" }
-serde.workspace = true
-serde_json.workspace = true
-toml = "0.8"
-
-[dev-dependencies]
-tempfile = "3"
